requirements.txt
langchain
python-dotenv
chainlit
langchain-openai
langchain_community